+++ package.json
@@ -1,8 +1,9 @@
 {
   "name": "askit",
+  "version": "0.0.0",
   "type": "module",
   "scripts": {
-    "dev": "dotenv -e .env tsx watch ./api/index.ts"
+    "dev": "dotenv -e .env tsx watch ./src/index.ts"
   },
   "dependencies": {
     "@slack/bolt": "^3.14.0",
